requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
bs4~=0.0.1
aiohttp~=3.9.5
beautifulsoup4~=4.12.2
pandas
jsonschema
pyyaml
ruamel~0.18.6
defusedxml
xmltodict
more_itertools
